soundcloud url can now be passed through via application scope
An observer has been placed on the plangular attribute, which now checks for the existence for src. If it does not exist will now get the soundcloud source from the application attribute, which is able to handle the value coming from the application scope. A demo will still have to be made showing this